# doc/api-doc.yaml
openapi: 3.0.3
info:
  title: rezeptbuch API
  version: "0.2.0"
  description: >
    REST API for managing recipes, their shared ingredient/tag/allergen
    master data, and cross-domain recipe search. A recipe has a title, a
    description, an origin, an ordered list of worksteps, an ordered list of
    ingredient lines (each pointing at a shared, reusable ingredient), and a
    set of tags. All validation is performed by the backend. Write endpoints
    (POST/PATCH/PUT/DELETE) are rate limited; GET endpoints and /health are
    exempt.
servers:
  - url: http://localhost:5000/api
    description: Local backend (direct)
  - url: http://localhost:8080/api
    description: Local frontend (nginx proxy)

paths:
  /health:
    get:
      summary: Liveness probe
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /recipes:
    get:
      summary: List recipes (newest first)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paginated recipes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipePage"
    post:
      summary: Create a recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeInput"
      responses:
        "201":
          description: Recipe created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "415":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"

  /recipes/search:
    get:
      summary: Filter recipes by tag, ingredient, allergen, and/or free text
      description: >
        All supplied query parameters combine with AND semantics (e.g.
        `?tag=breakfast&ingredient=flour` returns recipes matching both).
      parameters:
        - name: tag
          in: query
          schema:
            type: string
        - name: ingredient
          in: query
          schema:
            type: string
        - name: allergen
          in: query
          schema:
            type: string
        - name: q
          in: query
          description: Free-text match against title/description.
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paginated matching recipes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipePage"

  /recipes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get a recipe by id
      responses:
        "200":
          description: The recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      summary: Partially update a recipe
      description: >
        Only top-level fields present in the request body are updated.
        For `ingredients`/`worksteps`/`tags`, "present" is collection-level:
        supplying the field still replaces the whole collection wholesale,
        but omitting it leaves the existing collection untouched.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipePatch"
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      summary: Delete a recipe
      responses:
        "204":
          description: Deleted (no content)
        "404":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/RateLimited"

  /ingredients:
    get:
      summary: List ingredients (alphabetical), optionally prefix-filtered
      parameters:
        - name: q
          in: query
          description: Case-insensitive name-prefix filter, for autocomplete.
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paginated ingredients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngredientPage"
    post:
      summary: Create an ingredient master row
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngredientInput"
      responses:
        "201":
          description: Ingredient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"

  /ingredients/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get an ingredient by id
      responses:
        "200":
          description: The ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
        "404":
          $ref: "#/components/responses/Error"
    put:
      summary: Replace an ingredient's name/description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngredientInput"
      responses:
        "200":
          description: Updated ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      summary: Delete an ingredient
      description: Fails with 409 if the ingredient is used by any recipe.
      responses:
        "204":
          description: Deleted (no content)
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/RateLimited"

  /tags:
    get:
      summary: List tags (alphabetical), optionally prefix-filtered
      parameters:
        - name: q
          in: query
          description: Case-insensitive name-prefix filter, for autocomplete.
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paginated tags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagPage"
    post:
      summary: Create a tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
      responses:
        "201":
          description: Tag created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"

  /tags/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      summary: Delete a tag
      description: Renaming a tag is delete+recreate; there is no PUT/PATCH.
      responses:
        "204":
          description: Deleted (no content)
        "404":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/RateLimited"

  /allergens:
    get:
      summary: List allergens (alphabetical)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paginated allergens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllergenPage"
    post:
      summary: Create an allergen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllergenInput"
      responses:
        "201":
          description: Allergen created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allergen"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"

  /allergens/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      summary: Delete an allergen
      responses:
        "204":
          description: Deleted (no content)
        "404":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/RateLimited"

components:
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  schemas:
    Workstep:
      type: object
      required: [title, description]
      properties:
        step_number:
          type: integer
          readOnly: true
        title:
          type: string
          example: "Mix"
        description:
          type: string
          example: "Combine dry ingredients."

    RecipeIngredientInput:
      type: object
      description: >
        Either `ingredient_id` (reuse an existing master ingredient) or
        `name` (resolve-or-create by name) must be supplied.
      properties:
        ingredient_id:
          type: string
          format: uuid
        name:
          type: string
          example: "Baking powder"
        amount:
          type: string
          example: "200"
        unit:
          type: string
          example: "g"

    RecipeIngredient:
      type: object
      properties:
        ingredient_id:
          type: string
          format: uuid
        name:
          type: string
        amount:
          type: string
        unit:
          type: string
        position:
          type: integer

    RecipeTag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string

    RecipeInput:
      type: object
      required: [title, ingredients, worksteps]
      properties:
        title:
          type: string
          example: "Pancakes"
        description:
          type: string
          example: "Fluffy breakfast pancakes"
        origin:
          type: string
          nullable: true
          example: "US"
        tags:
          type: array
          items:
            type: string
          example: ["breakfast", "quick"]
        ingredients:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/RecipeIngredientInput"
        worksteps:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Workstep"

    RecipePatch:
      type: object
      description: Same shape as RecipeInput, but every field is optional.
      properties:
        title:
          type: string
        description:
          type: string
        origin:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        ingredients:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/RecipeIngredientInput"
        worksteps:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Workstep"

    Recipe:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        origin:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        worksteps:
          type: array
          items:
            $ref: "#/components/schemas/Workstep"
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/RecipeTag"

    RecipePage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Recipe"

    IngredientInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Flour"
        description:
          type: string
          nullable: true
          example: "All-purpose wheat flour"

    Ingredient:
      allOf:
        - $ref: "#/components/schemas/IngredientInput"
        - type: object
          properties:
            id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time

    IngredientPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Ingredient"

    TagInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "breakfast"

    Tag:
      allOf:
        - $ref: "#/components/schemas/TagInput"
        - type: object
          properties:
            id:
              type: string
              format: uuid

    TagPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Tag"

    AllergenInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "gluten"

    Allergen:
      allOf:
        - $ref: "#/components/schemas/AllergenInput"
        - type: object
          properties:
            id:
              type: string
              format: uuid

    AllergenPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Allergen"

    PageEnvelope:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true

  responses:
    Error:
      description: Error with a human-readable message
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Request failed validation; details map fields to messages
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Validation failed."
            details:
              title: ["Title is required."]
    RateLimited:
      description: Write rate limit exceeded (60 requests/minute per IP)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Rate limit exceeded."
            details: {}